% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{GetCellTypeSpecificTRN}
\alias{GetCellTypeSpecificTRN}
\title{Function to get cell types specific network}
\usage{
GetCellTypeSpecificTRN(tmp.seu, tmp.ident, gene.TF, tmp.phenotype)
}
\arguments{
\item{tmp.seu}{input Seurat v3 project}

\item{tmp.ident}{input celltypes}

\item{gene.TF}{require TF list, default is downloaded from AnimalTFDB}

\item{tmp.phenotype}{Moran'I, you can obtained this by run \link{GetVariableGeneInTrajectory}}
}
\value{
a tna object defined by RTN package, which contains master regulator and transcription factor regulatory network information
}
\description{
This function will genereate cell type specific network by RTN package, the TRN network reconstruction require zero-conserved imputation method ALRA implemented by Seurat
}
\examples{
GetCellTypeSpecificTRN(tmp.seu = seu, tmp.ident = "2C_like", gene.TF = gene.TF, tmp.phenotype = tmp.phenotype)
}
