% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{GetVariableGeneInTrajectory}
\alias{GetVariableGeneInTrajectory}
\title{Function to get variable gene in trajectory}
\usage{
GetVariableGeneInTrajectory(
  tmp.cds,
  tmp.cores = 8,
  tmp.neighbor_graph = "knn",
  tmp.q_value_cutoff = 1e-08
)
}
\arguments{
\item{tmp.cds}{monocle3 cds object}

\item{tmp.cores}{the number of cores to be used while testing each gene for differential expression, default is 8}

\item{tmp.neighbor_graph}{String indicating what neighbor graph to use. "principal_graph" and "knn" are supported. Default is "knn", but "principal_graph" is recommended for trajectory analysis.}

\item{tmp.q_value_cutoff}{cutoff of q value, default is 1e-8}
}
\value{
a named vector with Moran's I as a phenotype for the transcriptional regulatory network reconstruction, names as genes
}
\description{
This function will get the variable gene along the trajectory, by Test genes for differential expression based on the low dimensional embedding and the principal graph
}
\examples{
tmp.phenotype <- GetVariableGeneInTrajectory(tmp.cds = cds)
}
