% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MinMaxScale}
\alias{MinMaxScale}
\title{Function to peform min max scale}
\usage{
MinMaxScale(x)
}
\arguments{
\item{x}{input vector}
}
\value{
a vector in the interval of 0 between 1
}
\description{
This function perform min max scale
}
\examples{
MinMaxScale(rnorm(100))
}
