% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myRemoveNA}
\alias{myRemoveNA}
\title{Function to remove NA}
\usage{
myRemoveNA(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
a data.frame without NA
}
\description{
This function replace NA by 0
}
\examples{
myRemoveNA(data.frame(x = c(0,1,NA),y = c(1,2,3)))
}
