% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{RunMonocle3}
\alias{RunMonocle3}
\title{Function to run monocle3 pipeline}
\usage{
RunMonocle3(tmp.seu, tmp.assay = "integrated", tmp.minimal_branch_len = 5)
}
\arguments{
\item{tmp.seu}{input seurat v3 object}

\item{tmp.assay}{which assay to use from seurat v3 object, default is integrated}

\item{tmp.minimal_branch_len}{The minimal length of the diameter path for a branch to be preserved during graph pruning procedure. Default is 5.}
}
\value{
A monocle3 cds object with trajectory information
}
\description{
This function will run monocle3 pipeline from seurat object
}
\examples{
cds <- RunMonocle3(tmp.seu = tmp.seu, tmp.minimal_branch_len = 5)
}
